added mirror button - not working
