+++ src/lib/components/CaseCard.svelte
@@ -1,50 +1,57 @@
 <script lang="ts">
-        import TwistyPlayer from '$lib/components/TwistyPlayer.svelte';
-        import { casesStatic } from '$lib/casesStatic';
-        import { casesState } from '$lib/casesState.svelte';
-        import { globalState } from '$lib/globalState.svelte';
-        import getRotationAlg from '$lib/rotation';
-        import getStickeringString from '$lib/stickering';
-        import type { CaseId, GroupId } from '$lib/types/group';
+	import TwistyPlayer from '$lib/components/TwistyPlayer.svelte';
+	import { casesStatic } from '$lib/casesStatic';
+	import { casesState } from '$lib/casesState.svelte';
+	import { globalState } from '$lib/globalState.svelte';
+	import getRotationAlg from '$lib/rotation';
+	import getStickeringString from '$lib/stickering';
+	import type { CaseId, GroupId } from '$lib/types/group';
+	import { Button } from 'flowbite-svelte';
 
-        let {
-                groupId,
-                caseId
-        }: {
-                groupId: GroupId;
-                caseId: CaseId;
-        } = $props();
+	let {
+		groupId,
+		caseId
+	}: {
+		groupId: GroupId;
+		caseId: CaseId;
+	} = $props();
 
-        const staticData = casesStatic[groupId][caseId];
-        const caseState = casesState[groupId][caseId];
+	const staticData = casesStatic[groupId][caseId];
+	const caseState = casesState[groupId][caseId];
 
-        let mirrored = $state(false);
+	const selectedAlgorithm = $derived(
+		caseState.customAlgorithm.left.trim() ||
+			staticData.algPool[caseState.algorithmSelection.right] ||
+			staticData.algPool[0] ||
+			''
+	);
 
-        const selectedAlgorithm = $derived(
-                caseState.customAlgorithm.left.trim() ||
-                        staticData.algPool[caseState.algorithmSelection.left] ||
-                        staticData.algPool[0] ||
-                        ''
-        );
+	const selectedScramble = $derived(
+		staticData.scramblePool[caseState.algorithmSelection.right] || staticData.scramblePool[0] || ''
+	);
 
-        const selectedScramble = $derived(
-                staticData.scramblePool[caseState.algorithmSelection.left] ||
-                        staticData.scramblePool[0] ||
-                        ''
-        );
-
-        let stickeringString = $derived(
-                getStickeringString(
-                        globalState.crossColor,
-                        globalState.frontColor,
-                        staticData.pieceToHide,
-                        mirrored
-                )
-        );
-        const setupRotation = $derived(getRotationAlg(globalState.crossColor, globalState.frontColor));
+	let stickeringString = $derived(
+		getStickeringString(
+			globalState.crossColor,
+			globalState.frontColor,
+			staticData.pieceToHide,
+			caseState.mirrored
+		)
+	);
+	const setupRotation = $derived(getRotationAlg(globalState.crossColor, globalState.frontColor));
 </script>
 
 <div class="flex items-center rounded-2xl border-3">
-        <TwistyPlayer alg={selectedAlgorithm} {setupRotation} setupAlg={selectedScramble} {stickeringString} />
-        <span> {selectedAlgorithm} </span>
+	<TwistyPlayer
+		alg={selectedAlgorithm}
+		{setupRotation}
+		setupAlg={selectedScramble}
+		{stickeringString}
+	/>
+	<span> {selectedAlgorithm} </span>
+	<Button
+		onclick={() => {
+			caseState.mirrored = !caseState.mirrored;
+		}}>Mirror</Button
+	>
 </div>
